refactor(kitchenpos): segment model and repository across domain layer
