+++ src/main/java/kitchenpos/menu/infra/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.infra;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.model.MenuGroup;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
