+++ src/main/java/kitchenpos/order/domain/model/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.model;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
