+++ src/main/java/kitchenpos/order/domain/repository/OrderTableRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.repository;
+
+import kitchenpos.order.domain.model.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
