+++ src/main/java/kitchenpos/order/infra/JpaOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.infra;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.model.Order;
+import kitchenpos.order.domain.repository.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
