+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,11 +1,11 @@
 package kitchenpos.product.application;
 
 import kitchenpos.common.infra.PurgomalumClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.menu.domain.model.Menu;
+import kitchenpos.menu.domain.model.MenuProduct;
+import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
