+++ src/test/java/kitchenpos/menu/application/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.application;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.model.MenuGroup;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
 
 import java.util.*;
