+++ src/test/java/kitchenpos/menu/application/InMemoryMenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.application;
 
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.model.Menu;
+import kitchenpos.menu.domain.repository.MenuRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
