+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,12 +3,12 @@ package kitchenpos.menu.application;
 import kitchenpos.common.infra.FakePurgomalumClient;
 import kitchenpos.product.application.InMemoryProductRepository;
 import kitchenpos.common.infra.PurgomalumClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.menu.domain.model.Menu;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.model.MenuProduct;
+import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
