+++ src/test/java/kitchenpos/order/application/InMemoryOrderRepository.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.application;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.model.Order;
+import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.domain.model.OrderStatus;
+import kitchenpos.order.domain.model.OrderTable;
 
 import java.util.*;
