+++ src/test/java/kitchenpos/order/application/OrderTableServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.application;
 
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.domain.model.OrderStatus;
+import kitchenpos.order.domain.model.OrderTable;
+import kitchenpos.order.domain.repository.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
